app/api/pyproject.toml
[project]
name = "ticketing-api"
version = "0.1.0"
description = "Ticketing system — API service"
requires-python = ">=3.13,<3.14"

dependencies = [
    # Web framework
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",

    # Settings and validation
    "pydantic[email]==2.10.4",
    "pydantic-settings==2.7.1",

    # Database — PostgreSQL via SQLAlchemy 2 async
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.31.0",
    "alembic==1.14.0",

    # Azure SDKs (passwordless auth via Workload Identity)
    "azure-identity==1.21.0",
    "azure-keyvault-secrets==4.9.0",
    "azure-servicebus==7.13.0",

    # Redis
    "redis[hiredis]==5.2.1",

    # Observability
    "azure-monitor-opentelemetry==1.6.5",
    "opentelemetry-instrumentation-fastapi==0.50b0",
    "opentelemetry-instrumentation-sqlalchemy==0.50b0",
    "opentelemetry-instrumentation-redis==0.50b0",
    "structlog==24.4.0",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "pytest-cov==6.0.0",
    "httpx==0.28.1",                    # used by FastAPI's TestClient
    "ruff==0.8.4",
    "mypy==1.13.0",
    "types-redis==4.6.0.20241004",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ticketing_api"]

# ── Ruff (linter + formatter) ─────────────────────────────────────────────────

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # flake8-bandit (security)
    "ASYNC", # flake8-async
]
ignore = [
    "S101",  # use of `assert` — fine in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S"]   # security checks too noisy in tests

# ── Mypy ──────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]
disallow_any_explicit = false
warn_return_any = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ── Pytest ────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "integration: integration tests requiring docker-compose stack",
]